chore: rename app entry point to make it distinct from the main css file
