+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./index";
+import App from "./app";
 import "./app.css";
 import "./i18n";
 import { analytics } from "./services/analyticsService";
